Code refactoring 2 (#341)
* :lipstick: pump up the version number

* :tractor: further code refactoring
