+++ .moban.cd/moban.yml
@@ -4,8 +4,8 @@ organisation: moremoban
 author: C. W.
 contact: wangc_2011@hotmail.com
 license: MIT
-version: 0.6.4
-current_version: 0.6.4
+version: 0.6.5
+current_version: 0.6.5
 release: 0.6.4
 branch: master
 master: index
