+++ docs/conf.py
@@ -25,7 +25,7 @@ project = 'moban'
 copyright = '2017-2019 Onni Software Ltd.'
 author = 'C. W.'
 # The short X.Y version
-version = '0.6.4'
+version = '0.6.5'
 # The full version, including alpha/beta/rc tags
 release = '0.6.4'
