+++ moban/_version.py
@@ -1,2 +1,2 @@
-__version__ = "0.6.4"
+__version__ = "0.6.5"
 __author__ = "C. W."
