+++ moban/core/mobanfile/__init__.py
@@ -10,12 +10,12 @@ from moban.externals import reporter
 from moban.core.utils import verify_the_existence_of_directories
 from moban.deprecated import handle_copy, handle_requires
 from moban.core.data_loader import merge
-from moban.mobanfile.targets import (
+from moban.core.moban_factory import expand_template_directories
+from moban.core.mobanfile.targets import (
     parse_targets,
     extract_target,
     extract_group_targets,
 )
-from moban.core.moban_factory import expand_template_directories
 
 LOG = logging.getLogger(__name__)
