+++ tests/mobanfile/test_targets.py
@@ -2,8 +2,8 @@ import uuid
 
 import fs.path
 from nose.tools import eq_, raises
-from moban.mobanfile import targets
 from moban.exceptions import GroupTargetNotFound
+from moban.core.mobanfile import targets
 from moban.core.definitions import TemplateTarget
 
 TEMPLATE = "a.jj2"
