+++ tests/mobanfile/test_templates.py
@@ -1,7 +1,7 @@
 import fs.path
 from mock import patch
 from nose.tools import eq_
-from moban.mobanfile.templates import handle_template
+from moban.core.mobanfile.templates import handle_template
 
 
 class TestHandleTemplateFunction:
